requirements.txt
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34


Adam benchmark=cifar
    0.005 3
    Final Average accuracy = 56.64, forget = 0.10768421052631581
    0.001 3
    Final Average accuracy = 57.92, forget = 0.12610526315789472
    0.001 5
    Final Average accuracy = 61.53000000000001, forget = 0.0968421052631579
    0.001 7
    Final Average accuracy = 58.239999999999995, forget = 0.09305263157894737


Adam benchmark=imagenet
    0.005 1
    Final Average accuracy = 49.949999999999996, forget = 0.10429824561403508
    0.001 1
    Final Average accuracy = 50.39166666666667, forget = 0.1432456140350877
    0.001 3
    Final Average accuracy = 52.375, forget = 0.10938596491228068
    0.001 5
    Final Average accuracy = 49.74166666666666, forget = 0.08035087719298245


Adam benchmark=cub
    5e-05 7
    Final Average accuracy = 56.42158266499175, forget = 0.09116190149802149
    0.0001 7
    Final Average accuracy = 57.968492135374504, forget = 0.10521606543233256
    0.0001 5
    Final Average accuracy = 57.41270535240684, forget = 0.11385536313772711
    2.5e-05 7
    Final Average accuracy = 56.64637057471565, forget = 0.08859017707445614






RMSProp benchmark=cifar
    0.005 3
    Final Average accuracy = 56.71, forget = 0.09042105263157897
    0.0025 1
    Final Average accuracy = 57.540000000000006, forget = 0.11726315789473685
    0.0025 3
    Final Average accuracy = 58.85999999999999, forget = 0.0987368421052632


RMSProp benchmark=imagenet
    0.005 1
    Final Average accuracy = 50.50833333333333, forget = 0.10868421052631579
    0.0025 3
    Final Average accuracy = 49.80833333333332, forget = 0.08035087719298245


RMSProp benchmark=cub
    0.0001 3
    Final Average accuracy = 59.619074793566405, forget = 0.12699524402730242
    0.0001 5
    Final Average accuracy = 59.16213010931872, forget = 0.10688729265270944
    5e-05 5
    Final Average accuracy = 60.703496487262555, forget = 0.11523337514959062
    5e-05 3
    Final Average accuracy = 58.94207558082441, forget = 0.13088286185650286
    2.5e-05 3
    Final Average accuracy = 59.04056733089932, forget = 0.12719170494516777
    2.5e-05 5
    Final Average accuracy = 60.4798966287333, forget = 0.1189558596962469
    1e-05 5
    Final Average accuracy = 60.5492951888967, forget = 0.11698810763814553









Adagrad benchmark=imagenet
    0.001 3
    Final Average accuracy = 54.34166666666666, forget = 0.05956140350877195
    0.001 5
    Final Average accuracy = 57.01666666666667, forget = 0.05175438596491229
    0.001 7
    Final Average accuracy = 55.0, forget = 0.04798245614035089
    0.005 1
    Final Average accuracy = 55.55833333333332, forget = 0.13043859649122808
    0.005 3
    Final Average accuracy = 55.15833333333334, forget = 0.11517543859649121
    0.0025 3
    Final Average accuracy = 57.78333333333334, forget = 0.08263157894736843
    0.0025 5
    Final Average accuracy = 56.63333333333334, forget = 0.08254385964912282
    0.0005 7
    Final Average accuracy = 55.133333333333326, forget = 0.0387719298245614


benchmark=imagenet
Parameters: (0.01,0.9,0.0)     44.7
Parameters: (0.01,0.8,0.0)     43.8
Parameters: (0.01,0.9,0.1)     44.1
Parameters: (0.01,0.8,0.1)     44.2
Parameters: (0.005,0.9,0.0)     42.6
Parameters: (0.005,0.8,0.0)     42.1

benchmark=cub
Parameters: (0.005,0.9,0.0)    28.1
Parameters: (0.0025,0.9,0.0)    25.9
